gateway and docker-compose

+++ users/schema/20240502201440_user_table.up.sql
@@ -1,12 +1,12 @@
 create table if not exists users(
-  id integer not null unique,
+  id bigint not null unique,
   firstname varchar not null,
   username varchar,
   createdat date not null
 );
 
 create table if not exists game(
-  ownerId integer not null,
+  ownerId bigint not null,
   score integer DEFAULT 0,
   gasStorage integer DEFAULT 1,
   gasMining integer DEFAULT 1,
@@ -15,8 +15,8 @@ create table if not exists game(
 );
 
 create table if not exists userReferral(
-  inviterId integer not null,
-  referralId integer not null,
+  inviterId bigint not null,
+  referralId bigint not null,
   CONSTRAINT fk_userReferral_inviterId FOREIGN KEY(inviterId) REFERENCES users(id),
   CONSTRAINT fk_userReferral_referralId FOREIGN KEY(referralId) REFERENCES users(id)
 );
